auth_key1 fix; todo server_DH_params_ok

+++ essential/serialize.h
@@ -1,7 +1,7 @@
 #ifndef SERIALIZE_H
 #define SERIALIZE_H
 #include "buf.h"
-buf_t serialize_bytes(uint8_t *bytes, size_t size);
+buf_t serialize_bytes(uint8_t *bytes, uint32_t size);
 buf_t serialize_string(const char *string);
 buf_t serialize_str(buf_t b);
 #endif /* ifndef SERIALIZE_H */
